Separate :calculator logic from date/time types
